feat: getChoseng을 utils에서 별도 함수로 분리합니다. (#192)
* getChoseong분리

* write test code

* getChoseong import

* remove useless import statemenet

* remove unused file
