+++ docs/src/pages/docs/api/getChoseong.en.mdx
@@ -1,0 +1,19 @@
+# getChoseong
+
+
+Extracts the Choseong from a Korean word. (Example: 사과 -> 'ㅅㄱ')
+
+```typescript
+function getChoseong(
+  // Korean string from which to extract the choseong 
+  word: string
+): string;
+```
+
+## Examples
+
+```tsx
+getChoseong('사과') // 'ㅅㄱ'
+getChoseong('리액트') // 'ㄹㅇㅌ'
+getChoseong('띄어 쓰기') // 'ㄸㅇ ㅆㄱ'
+```
