+++ docs/src/pages/docs/api/getChoseong.ko.mdx
@@ -1,0 +1,20 @@
+# getChoseong
+
+단어에서 초성을 추출합니다. (예: `사과` -> `'ㅅㄱ'`)
+
+자세한 예시는 아래 Example을 참고하세요.
+
+```typescript
+function getChoseong(
+  // 초성을 추출 할 한글 문자열
+  word: string
+): string;
+```
+
+## Examples
+
+```tsx
+getChoseong('사과') // 'ㅅㄱ'
+getChoseong('리액트') // 'ㄹㅇㅌ'
+getChoseong('띄어 쓰기') // 'ㄸㅇ ㅆㄱ'
+```
